expandablelistview navigation drawer all with cursors

+++ thoth/src/main/java/com/concentricsky/android/thoth/ThothDatabaseHelper.java
@@ -91,7 +91,7 @@ public class ThothDatabaseHelper
         return c;
     }
 
-    public Cursor getFeedCursor(int tag_id)
+    public Cursor getFeedCursor(long tag_id)
     {
         SQLiteDatabase db = mOpenHelper.getReadableDatabase();
         Cursor c = db.rawQuery("SELECT _id,title FROM "+Feed.FEED_TABLE_NAME+" JOIN "+Feed.FEEDTAG_TABLE_NAME+" ON feed_id=_id WHERE tag_id=?", new String[] {String.valueOf(tag_id)});
